update tăng giảm quantity khi mượn và trả sách
